Add function to translate .envrc files
Some projects include a .envrc file to export the default environment variables
that the project relies on. Since fish does not export variables in the
same was as bash, create a quick and dirty function to translate the
envrc file.

Make sure we trust the input since we are using eval.