Add function to activate virtual env based on pwd
Currently storing virtual envs in the ~/.venv directory. Instead of
having to source the whole virtual env which involves typing the whole
path, we can lean on some fish magic to construct the virtual env path
for us.

This assumes that the virtual env we have created is named for the
project directory. If it is all set up correctly, then simply calling
venvup will activate our virtual environment for us.